Match example on path, not file
The x-api matches the example based on the entire path, not on just
the file name. This makes it possible to create more idiomatic file
structures.

In particular, in Java, if the example is a directory at the same level
as main or test, then people can configure their IDE to load
./src/example into their classpath when they load the project with
gradle.