Add configurable solution pattern
Previously the pattern used to check if an example solution was present
for a file was hardcoded to `[Ee]xample`.

Since support was added for ignored `.meta` directory it is no longer
necessary for a solution filename to contain `example` to avoid being
served to the student.

This patch adds support for configuring the pattern used to detect the
presence of an example solution.

It is set via the optional key `solution_pattern` in the track's
`config.json` file. The value is a string which is used to create a
regular expression.

If the key is not present the default value of `[Ee]xample` will be
used.