Add fmt command
Much like Go's very own fmt command, this will consistently format a
track's configuration files, currently the `config.json` and
`maintainers.json` files. This will avoid any mixed indentation or
whitespace issues, and ensure correct JSON syntax, that is as pretty as
can be. It also allows for further formatting to the contents of
these files. For the `config.json` file the fmt command additionally
sorts each exercise's list of topics and normalizes the topic name.

Closes #18