Add a snake case version of the slug (#89)
Tracks should be able to generate the expected filename for an exercise,
in order to provide the exact command that people need to use to run
the tests for an exercise.

The typical naming conventions are:

- `kebab-case` (e.g. JavaScript)
- `snake_case` (e.g. Ruby)
- `MixedCase` (e.g. Groovy)

We already have support for `kebab-case` (the slug itself)
and `MixedCase`. This adds the missing `snake_case`.