Adjust fixtures to match expected tree
The config files that get reformatted live in the following
directory structure:

    a-track/
    ├── config
    │   └── maintainers.json
    └── config.json

The fixtures had both the track config and the maintainer
config at the same level, which is making it tricky to
properly test the fmt command.

This sets us up for a refactoring that will let us write better
tests.