Get rid of now-spurious return value
The formatFile function both reads from a file and writes to a file.
By default it reads and writes to the same file, but it's
configurable so that the tests can write to a temporary file.

Now that the tests are checking the actual written value, we can get rid
of this left-over return value.