Fix conceptual shear in fmt command implementation
The tests are stripping off the trailing newline from the config
file that gets read in, but the implementation leaves it in.

This causes the diff to contain a line that is not found in the
tests.

This will all eventually go away, but I need to normalize it
before I can do a bigger refactoring.