Only create directory if necessary
The fmt command can write to a separate directory rather than writing to
the file that it read from.

This is handy for the tests, which need to be able to check the
serialization without clobbering the existing fixtures.

This ensures that we don't create directories unnecessarily. The tests
are in charge of cleaning up any directories that get created.