config: Only marshal Slug, UUID, Deprecated of deprecated exercises
We don't need any of the other fields, and it's deceptive to keep them
in the JSON file because a maintainer may mistakenly believe them to
be significant, but in fact they are not; the website does not display
any of the removed fields:
https://github.com/exercism/website/blob/b3314e84c52b52cf5f253386c3f0bf4127ca09aa/app/views/my/tracks/_show_side_exercises.html.haml#L26-L34

Two new tests were added: TestMarshalActive and TestMarshalDeprecated.
TestMarshalActive passes both before and after the associated code
change.
TestMarshalDeprecated fails before the associated code change, and
passes after.

Closes https://github.com/exercism/configlet/issues/140